:recycle: refactor: move package and change name company_list_page
