+++ lib/features/dashboard/data/model/company_model.dart
