+++ lib/features/dashboard/domain/entities/company.dart
