+++ test/features/company/domain/usecase/get_companies_use_case_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:tractian_challenge_one/features/company/domain/repository/company_repository.dart';
-import 'package:tractian_challenge_one/features/company/domain/usecase/get_companies_use_case.dart';
+import 'package:tractian_challenge_one/features/dashboard/domain/repository/company_repository.dart';
+import 'package:tractian_challenge_one/features/dashboard/domain/usecase/get_companies_use_case.dart';
 
 import '../utils/class_mock.dart';
 import '../../stubs.dart';
