+++ test/features/company/presentation/widgets/company_list_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
-import 'package:tractian_challenge_one/features/company/domain/entities/company.dart';
-import 'package:tractian_challenge_one/features/company/presentation/widgets/company_list.dart';
+import 'package:tractian_challenge_one/features/dashboard/domain/entities/company.dart';
+import 'package:tractian_challenge_one/features/dashboard/presentation/widgets/company_list.dart';
 
 import '../../stubs.dart';
